+++ src/Products/PloneMeeting/browser/batchactions.py
@@ -19,6 +19,7 @@ from Products.CMFCore.utils import _checkPermission
 from Products.PloneMeeting import logger
 from Products.PloneMeeting.config import NO_COMMITTEE
 from Products.PloneMeeting.config import PMMessageFactory as _
+from Products.PloneMeeting.ftw_labels.utils import filter_access_global_labels
 from Products.PloneMeeting.utils import displaying_available_items
 from z3c.form.field import Fields
 from zope import schema
@@ -269,7 +270,10 @@ class PMLabelsBatchActionForm(LabelsBatchActionForm):
         """Only available when labels are enabled."""
         tool = api.portal.get_tool('portal_plonemeeting')
         cfg = tool.getMeetingConfig(self.context)
-        return cfg.getEnableLabels()
+        return 'labels' in cfg.getUsedItemAttributes()
+
+    def _filter_labels_vocabulary(self, jar):
+        return filter_access_global_labels(jar, mode='edit')
 
     def _can_change_labels(self):
         view = None