+++ CHANGES.rst
@@ -12,6 +12,12 @@ Changelog
 - Display `voting group` in `MeetingConfig.orderedContacts` so we can detect
   easily for which held position it is configured.
   [gbastien]
+- Added `MeetingConfig.labelsConfig` to be able to configure who can
+  view/edit labels.
+  [gbastien]
+- Added parameter `extra_ctx={}` to `utils._base_extra_expr_ctx` to be able
+  to pass extra context immediatelly without having to `dict.update` after.
+  [gbastien]
 
 4.2.24.3 (2025-09-04)
 ---------------------